// src/app/Shared/models/cues/cue.ts
export class CueDTO {
  index = -1; //-1: not a hotcue, 0-7: hotcue 1-8
  position: number; //cue position (in s?)
  name?: string; //AutoGrid / Beat Marker --> Grid (name autogenerated)
  type?: number; //0: Hotcue, 1: Fade In, 2: Fade-Out, 3: Load, 4: Grid, 5: Loop

  //Traktor
  order = 0; //Always 0... The display order in the screen for when they overlap?
  repeats = -1; //Always -1... Should it refer to wether it's a loop or not?

  constructor(
    position: number,
    index?: number,
    name?: string,
    type?: number,
    order?: number,
    repeats?: number,
  ) {
    this.name = name;
    if (typeof index === 'number') this.index = index;
    this.type = type;
    this.position = position;
    if (order) this.order = order;
    if (repeats) this.repeats = repeats;
  }
}